# model/src/main/java/com/fsa/types/FlightSearchResponse.java
package com.fsa.types;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fsa.dto.Flight;
import java.util.ArrayList;
import java.util.List;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@JsonIgnoreProperties(ignoreUnknown = true)
@JsonInclude(JsonInclude.Include.NON_NULL)
public class FlightSearchResponse {

  List<Flight> flightLis = new ArrayList<>();
  Error error;
}